Fix poudriere build_fs_violation for net/samba*
- Bump PORTREVISION for package change

The tevent module is being used during build of net/samba*, therefore the
bytecode .pyc was generated. When running poudriere with -t flag, it checks
filesystem violations and raises the tevent module was used error. The fix is to
generate the bytecode and package it.

PR:		243331
Submitted by:	sunpoet (myself)
Approved by:	maintainer (timeout, 25 days)