Update to the 20201227 snapshot of GCC 11.0.0.
This changes the processor defaults for 32- und 64-bit powerpc to
PPC7450 and POWER8, respectively, per a request by pkubaj@ that I
pushed upstream. [1]

Enable the new powerpcle architecture which this snapshot brings in
via upstream, per a submission by pkubaj@. [1]

The 20201227 snapshot was broken on FreeBSD (and presumably other
non-mainstream platforms), so we had to skip it.

PR:		251670 [1]