Update to the 20210102 snapshot of GCC 10.2.1.
This brings two backports for the aarch64 backend and one for x86,
plus three for the Fortran front end.

Enable the new powerpcle architecture which this snapshot brings in
via upstream, per a submission by pkubaj@. [1]

This also should fix the build on aarch64 when clang is the bootstrap
compiler. [2]

PR:		251670 [1], 250932 [2]