net-mgmt/glpi-plugins-fusioninventory-server: update to v9.5.0.1.0
Realign this software version with the current GLPI and FusionInventoy
available in ports. The last time this port was properly update was
in 2016, and since then multiple version where released. Please refer
to the project changelog to see what's new:

https://github.com/fusioninventory/fusioninventory-for-glpi/releases

This version is compatible with GLPI 9.5 series (glpi-9.5.3,1 in ports)

I also take mainteneship for this port