Mark net-im/mtxclient unsupported on FreeBSD 11
This is a bit hairy, since 11.4 is supported until 2021-09-30. With
only one consumer in the tree, which has good alternatives, I don't think
it is worth the effort to chase.

At issue is SSL support: mtxclient needs SSL 1.1, which base doesn't
supply. However, it **also** needs to be compatible with the SSL used
in the rest of the stack (Qt in particular), so I can't push a ports
SSL here without also chasing that all over and potentially breaking
other parts of the stack.

No PORTREVISION bump since the package doesn't change, and it never built
on 11 anyway.