New port deskutils/kongress, a conference companion
Kongress is a conference companion application. Its purpose is to help
participants to organize their visit or online participation in a
conference. It runs well on a (Linux) mobile phone, but works fine
on a desktop or laptop machine.

Release notes were on the KDE-releases mailing list but I copied
them into the pkg-descr; they are also available at
	https://dimitris.cc/kde/2021/01/28/Kongress_1_0.html

There is one patch because I didn't like the Qt5::Test dependency
and it seems to be optional anyway, MR submitted upstream.