Update deskutils/nextcloudclient to 3.1.2
Thanks to Guido for providing the patch.

This version has changes to the auth code.
Guido added the -DNO_SHIBBOLETH=1 to disable code that fails to compile now.
It has been removed from the source code as dead code shortly after the release:
https://github.com/nextcloud/desktop/commit/db0f1e245df27e5673df0931da0e4e0f56e1c0b6
Before that upstream builds had this NO_SHIBBOLETH=1 flag turned on.
It now depends on websockets in place of webkit due to this.
Apart from this change this is a routine update.

PR:		253319
Submitted by:	Guido Falsi