Upgrade to version 3.3.0
There is one functional change with regard to -e of -f in BC_ENV_ARGS resp.
DC_ENV_ARGS: these do no longer make the program exit after processing the
commands passed that way (which prevented to use the environment variables
to set parameters for interactive sessions, unless -f- was passed as the
last argument on the command line).