Security update to gitlab-ce 13.9.2. Changelog: https://about.gitlab.com/releases/2021/03/04/security-release-gitlab-13-9-2-released/
This commit also enforces an older version of devel/rubygem-google-protobuf of version 3.14.0.
This also linked PR.
This ensures that users to not upgrade by accident to a version that is core dumping.
So it is wanted that this port is currently not buildable to protect users from an update.

If you want to this upgrade, wait till devel/rubygem-google-protobuf is fixed or downgrade it to 3.14.0.

PR:		254014 254010
Security:	8bf856ea-7df7-11eb-9aad-001b217b3468