Remove LLDB option to remove the dependency on python-2.7
This port has been on a list of deprecated ports with removal planned on
2021-06-23. Multiple other ports depend on this particular LLVM version
and were also marked for removal. This change allows to retain this port
and the ports that depend on it.

Newer LLDB versions are available in the base system or in later LLVM ports.

The pkg-descr mentions the removal of LLDB from this port as suggested by
Ed Maste.

Approved by:	emaste
Differential Revision:	D29342