sysutils/runj: new port
Summary:
runj is an experimental, proof-of-concept OCI-compatible runtime for
FreeBSD jails.

Important: runj is a proof-of-concept and the implementation has not
been evaluated for its security. Do not use runj on a production
system. Do not run workloads inside runj that rely on a secure
configuration. This is a personal project, not backed by the
author's employer.

WWW: https://github.com/samuelkarp/runj

Test Plan:
- needs poudriere + usage tests still
           - waiting on upstream to confirm if they want a port atm

Subscribers: mat

Differential Revision: https://reviews.freebsd.org/D29292