Welcome the new mail/spamassassin-devel port.
The reason for this port is due to the following statement in the
spamassassin 3.4.5 announcement:

Ongoing development on the 3.4 branch has ceased. All future releases
and bug fixes will be on the 4.0 series, unless a new security issue
is found that necessitates a 3.4.6 release.

Upstream is the apache/spamassassing github read-only mirror.